extract BundleFileChange into own module
