+++ reconcile/change_owners/change_owners.py
@@ -1,22 +1,22 @@
 import logging
 import sys
 import traceback
-from collections import defaultdict
 
 from reconcile import queries
 from reconcile.change_owners.approver import GqlApproverResolver
 from reconcile.change_owners.bundle import (
-    BundleFileType,
     FileDiffResolver,
     QontractServerFileDiffResolver,
 )
 from reconcile.change_owners.change_types import (
-    BundleFileChange,
     ChangeTypePriority,
     ChangeTypeProcessor,
-    create_bundle_file_change,
-    get_priority_for_changes,
     init_change_type_processors,
+)
+from reconcile.change_owners.changes import (
+    BundleFileChange,
+    fetch_bundle_changes,
+    get_priority_for_changes,
 )
 from reconcile.change_owners.decision import (
     ChangeDecision,
@@ -112,124 +112,6 @@ def fetch_change_type_processors(
     return list(
         init_change_type_processors(change_type_list, file_diff_resolver).values()
     )
-
-
-def fetch_bundle_changes(comparison_sha: str) -> list[BundleFileChange]:
-    """
-    reaches out to the qontract-server diff endpoint to find the files that
-    changed within two bundles (the current one representing the MR and the
-    explicitely passed comparision bundle - usually the state of the master branch).
-    """
-    changes = gql.get_diff(comparison_sha)
-    bundle_changes = _parse_bundle_changes(changes)
-    try:
-        return _aggregate_file_moves(bundle_changes)
-    except Exception as e:
-        logging.error(f"Failed to post process bundle changes: {e}")
-        # if any post processing action fails, we want the raw bundle changes
-        # to be reviewed
-        return bundle_changes
-
-
-def _aggregate_file_moves(
-    bundle_changes: list[BundleFileChange],
-) -> list[BundleFileChange]:
-    """
-    This function tries to detect file moves by looking at the bundle changes. If an
-    add and remove file change with the same content is detected, those changes are
-    replaced with a single move change where the only difference is the path.
-    """
-    move_candidates: dict[str, list[BundleFileChange]] = defaultdict(list)
-    new_bundle_changes = []
-    for c in bundle_changes:
-        potential_move_sha = None
-        if c.is_file_creation():
-            potential_move_sha = c.new_content_sha()
-        elif c.is_file_deletion():
-            potential_move_sha = c.old_content_sha()
-        if potential_move_sha:
-            move_candidates[potential_move_sha].append(c)
-        else:
-            new_bundle_changes.append(c)
-
-    for candidates in move_candidates.values():
-        if len(candidates) == 2:
-            # if there are two candidates, they could represent a file move.
-            # lets check if that is the case
-            deletion = (
-                candidates[0] if candidates[0].is_file_deletion() else candidates[1]
-            )
-            addition = (
-                candidates[0] if candidates[0].is_file_creation() else candidates[1]
-            )
-            if (
-                deletion != addition
-                and deletion.fileref.file_type == addition.fileref.file_type
-                and deletion.fileref.path != addition.fileref.path
-            ):
-                # the candidates represent a file move. let's add a new
-                # change that represents the move
-                move_change = create_bundle_file_change(
-                    deletion.fileref.path,
-                    deletion.fileref.schema,
-                    deletion.fileref.file_type,
-                    deletion.old,
-                    addition.new,
-                )
-                if move_change:
-                    # move_change will always be present. this check is just to
-                    # satisfy mypy
-                    new_bundle_changes.append(move_change)
-                    continue
-
-        # the candidates turned out to be unrelated changes. lets add them back
-        # to the list of changes
-        new_bundle_changes.extend(candidates)
-    return new_bundle_changes
-
-
-def _parse_bundle_changes(bundle_changes) -> list[BundleFileChange]:
-    """
-    parses the output of the qontract-server /diff endpoint
-    """
-    datafiles = bundle_changes["datafiles"].values()
-    resourcefiles = bundle_changes["resources"].values()
-    logging.debug(
-        f"bundle contains {len(datafiles)} changed datafiles and {len(resourcefiles)} changed resourcefiles"
-    )
-
-    change_list = []
-    for c in datafiles:
-        bc = create_bundle_file_change(
-            path=c.get("datafilepath"),
-            schema=c.get("datafileschema"),
-            file_type=BundleFileType.DATAFILE,
-            old_file_content=c.get("old"),
-            new_file_content=c.get("new"),
-        )
-        if bc is not None:
-            change_list.append(bc)
-        else:
-            logging.debug(
-                f"skipping datafile {c.get('datafilepath')} - no changes detected"
-            )
-
-    for c in resourcefiles:
-        bc = create_bundle_file_change(
-            path=c.get("resourcepath"),
-            schema=c.get("new", {}).get("$schema", c.get("old", {}).get("$schema")),
-            file_type=BundleFileType.RESOURCEFILE,
-            old_file_content=c.get("old", {}).get("content"),
-            new_file_content=c.get("new", {}).get("content"),
-        )
-        if bc is not None:
-            change_list.append(bc)
-        else:
-            logging.debug(
-                f"skipping resourcefile {c.get('resourcepath')} - no changes detected"
-            )
-
-    return change_list
 
 
 CHANGE_TYPE_PROCESSING_MODE_LIMITED = "limited"
