+++ reconcile/test/change_owners/test_change_type_coverage.py
@@ -5,8 +5,8 @@ from reconcile.change_owners.change_types import (
     Approver,
     ChangeTypeContext,
     DiffCoverage,
-    create_bundle_file_change,
 )
+from reconcile.change_owners.changes import create_bundle_file_change
 from reconcile.change_owners.diff import (
     Diff,
     DiffType,
